// src/redux/features/review/reviewRating.types.ts
export type Reviews = {
  id: number;
  item_id: {
    item_id: number;
    mainOwner: {
      id: number;
      name: string;
      username: string;
      email: string;
      user_type: string;
      user_photo: string;
      user_token: string;
      website: string;
      country: {
        country_id: number;
        country_name: string;
        vat_price: number;
        country_badges: string;
      };
      about: string;
      payment_detail: string;
      facebook_url: string;
      twitter_url: string;
      gplus_url: string;
      instagram_url: string;
      linkedin_url: string;
      pinterest_url: string;
      verified: number;
      user_document_verified: number;
      country_badge: number;
      drop_status: string;
      city: string;
      zip_code: number;
      address: string;
      mobile: string;
      first_name: string;
      last_name: string;
      organization: string;
      designation: string;
      experience: string;
      critic_organization: string;
      critic_designation: string;
      critic_experience: string;
      critic_award: string[];
      artist_award: string[];
      image_path: string;
      critic_document: string[];
      artist_document: string[];
      total_follower: number;
    };
    item_token: string;
    item_name: string;
    item_slug: string;
    item_desc: string;
    item_shortdesc: string;
    item_thumbnail: string;
    item_preview: string;
    item_file: string;
    item_category: {
      category_detaill: {
        cat_id: number;
        category_name: string;
        category_slug: string;
        category_status: number;
        menu_order: number;
        drop_status: string;
        image: string;
      };
    }[];
    regular_price: number;
    item_tags: string;
    item_featured: string;
    size_id: {
      id: number;
      name: string;
    };
    medium_id: {
      id: number;
      name: string;
    };
    material_id: {
      id: number;
      name: string;
    };
    subject_id: {
      id: number;
      name: string;
    };
    is_adult: number;
    avg_critic_rating: number;
    avg_customer_rating: number;
    total_favorite: number;
    current_owner_id: {
      id: number;
      name: string;
      username: string;
      email: string;
      user_type: string;
      user_photo: string;
      user_token: string;
      website: string;
      country: {
        country_id: number;
        country_name: string;
        vat_price: number;
        country_badges: string;
      };
      about: string;
      payment_detail: string;
      facebook_url: string;
      twitter_url: string;
      gplus_url: string;
      instagram_url: string;
      linkedin_url: string;
      pinterest_url: string;
      verified: number;
      user_document_verified: number;
      country_badge: number;
      drop_status: string;
      city: string;
      zip_code: number;
      address: string;
      mobile: string;
      first_name: string;
      last_name: string;
      organization: string;
      designation: string;
      experience: string;
      critic_organization: string;
      critic_designation: string;
      critic_experience: string;
      critic_award: string[];
      artist_award: string[];
      image_path: string;
      critic_document: string[];
      artist_document: string[];
      total_follower: number;
    };
    image_path: string;
    price_detail: {
      id: number;
      item_id: number;
      critic_avg_price: string;
      critic_rating_price: string;
      favorite_price: string;
      customer_rating_price: string;
      final_price: number;
      created_at: string;
      updated_at: string;
    };
    sale_status: number;
    total_avg_rating: number;
    total_review: number;
    review_opinion: number;
  };
  userDetail: {
    id: number;
    name: string;
    username: string;
    email: string;
    user_type: string;
    user_photo: string;
    user_token: string;
    website: string;
    country: {
      country_id: number;
      country_name: string;
      vat_price: number;
      country_badges: string;
    };
    about: string;
    payment_detail: string;
    facebook_url: string;
    twitter_url: string;
    gplus_url: string;
    instagram_url: string;
    linkedin_url: string;
    pinterest_url: string;
    verified: number;
    user_document_verified: number;
    country_badge: number;
    drop_status: string;
    city: string;
    zip_code: number;
    address: string;
    mobile: string;
    first_name: string;
    last_name: string;
    organization: string;
    designation: string;
    experience: string;
    critic_organization: string;
    critic_designation: string;
    critic_experience: string;
    critic_award: string[];
    artist_award: string[];
    image_path: string;
    critic_document: string[];
    artist_document: string[];
    total_follower: number;
  };
  review: string;
  rating: string;
  review_date: string;
  type: string;
};